backend/requirements.txt
fastapi==0.128.8
uvicorn==0.40.0
edge-tts==7.2.7
pydantic==2.12.5
python-multipart==0.0.20
edge-tts==7.2.7
pydantic==2.12.5
python-multipart==0.0.20
faiss-cpu>=1.13.0
pypdf==4.1.0
numpy
